requirements.txt
setuptools
pyrogram
asyncio
aiohttp
tgcrypto
requests
pytz
colorama
rich
python-dotenv